보다 구체화 / GetObjectAtGridPosition 함수가 3군데에 적용되어 있는데, 이에 static을 붙여서? 함수 참조를 할 수 있게 사용할까 고려필요

+++ Assets/Script/MakeACell.cs
@@ -25,11 +25,10 @@ public class MakeACell : MonoBehaviour {
 		
 	}
 
-    public void GenerateCell(Vector3 Where, int c)
+    public void GenerateCell(Vector3 Where, float c)
     {
         GameObject obj;
-        switch (a)
-        {
+        switch (a) {
             case 1:
                 obj = Instantiate(NormalCell, Where, transform.rotation);
                 obj.GetComponent<Cell>().value = c;
@@ -66,13 +65,8 @@ public class MakeACell : MonoBehaviour {
                 obj = Instantiate(LineGaroCell, Where, transform.rotation);
                 obj.GetComponent<Cell>().value = c;
                 break;
-
-
             default:
                 break;
-
-
-
         }
